// src/sanity/types.ts
export type HomePage = {
  _id: string;
  _type: "HomePage";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  HeroImage?: {
    asset?: {
      _ref: string;
      _type: "reference";
      _weak?: boolean;
      [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
    };
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    _type: "image";
  };
  HeroTitle?: string;
  CollectionOneTitle?: string;
  CollectionOneSubtitle?: string;
  CollectionOneShopifyCollection?: string;
  SectionOneTitle?: string;
  SectionOneSubtitle?: string;
  SectionOneShopifyCollection?: string;
  SectionOneCollectionDisplayName?: string;
  SectionOneImage?: {
    asset?: {
      _ref: string;
      _type: "reference";
      _weak?: boolean;
      [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
    };
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    _type: "image";
  };
  CategoryCollectionList?: Array<{
    _ref: string;
    _type: "reference";
    _weak?: boolean;
    _key: string;
    [internalGroqTypeReferenceTo]?: "NavigationCategory";
  }>;
  VideoSectionTitle?: string;
  VideoSectionLink?: string;
  VideoSectionButtonText?: string;
  CollectionTwoTitle?: string;
  CollectionTwoSubtitle?: string;
  CollectionTwoShopifyCollection?: string;
};

export type NavigationCategoryList = {
  _id: string;
  _type: "NavigationCategoryList";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  OrderedNavigationCategories?: Array<{
    _ref: string;
    _type: "reference";
    _weak?: boolean;
    _key: string;
    [internalGroqTypeReferenceTo]?: "NavigationCategory";
  }>;
};

export type NavCard = {
  _type: "NavCard";
  shopifyCategory?: string;
  title?: string;
  cardImage?: {
    asset?: {
      _ref: string;
      _type: "reference";
      _weak?: boolean;
      [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
    };
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    _type: "image";
  };
};

export type NavigationCategory = {
  _id: string;
  _type: "NavigationCategory";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  name?: string;
  slug?: Slug;
  Collections?: Array<{
    displayName?: string;
    shopifyCategory?: string;
    _type: "category-tuple";
    _key: string;
  }>;
  navigationCards?: Array<
    {
      _key: string;
    } & NavCard
  >;
};

export type SanityImageCrop = {
  _type: "sanity.imageCrop";
  top?: number;
  bottom?: number;
  left?: number;
  right?: number;
};

export type SanityImageHotspot = {
  _type: "sanity.imageHotspot";
  x?: number;
  y?: number;
  height?: number;
  width?: number;
};
// ...
export type Slug = {
  _type: "slug";
  current?: string;
  source?: string;
};

export type AllSanitySchemaTypes =
  | SanityImagePaletteSwatch
  | SanityImagePalette
  | SanityImageDimensions
  | SanityFileAsset
  | Geopoint
  | HomePage
  | NavigationCategoryList
  | NavCard
  | NavigationCategory
  | SanityImageCrop
  | SanityImageHotspot
  | SanityImageAsset
  | SanityAssetSourceData
  | SanityImageMetadata
  | Slug;
export declare const internalGroqTypeReferenceTo: unique symbol;